pyproject.toml
[tool.poetry]
name = "safety-report"
version = "0.1.0"
description = ""
authors = ["silofox <silofox@uisn.co.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymupdf = "^1.24.9"
pytesseract = "^0.3.10"
pillow = "^10.4.0"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
